Cleaned up other missing virtual dtors and unsigned int comparison warnings

+++ trunk/src/Polygon.cpp
@@ -37,9 +37,9 @@ ExecDrawOutline() const {
 
 void ol::Poly::
 SetVertex( int index, const Vec2D &newValue ) {
-   if( index < 0 || index >= data.GetVertices().size()) {
-      OlError( "Invalid vertex index: " + VarToString( index )
-         + " ( Number of vertices: " + VarToString( data.GetVertices().size()) + " )" );
+	if( index < 0 || index >= (int)data.GetVertices().size()) {
+		OlError( "Invalid vertex index: " + VarToString( index )
+		+ " ( Number of vertices: " + VarToString( (int)data.GetVertices().size()) + " )" );
    }
    
    data.GetVertices()[index] = newValue;
